sdk: Move timeline redaction test to new file
